Voip

+++ src/com/bsb/hike/db/HikeConversationsDatabase.java
@@ -1,7 +1,17 @@
 package com.bsb.hike.db;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.LinkedHashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Set;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -36,18 +46,41 @@ import com.bsb.hike.bots.BotInfo;
 import com.bsb.hike.bots.BotUtils;
 import com.bsb.hike.db.DBConstants.HIKE_CONV_DB;
 import com.bsb.hike.db.DatabaseErrorHandlers.CustomDatabaseErrorHandler;
-import com.bsb.hike.models.*;
+import com.bsb.hike.models.ContactInfo;
+import com.bsb.hike.models.ConvMessage;
 import com.bsb.hike.models.ConvMessage.ConvMessageComparator;
 import com.bsb.hike.models.ConvMessage.OriginType;
 import com.bsb.hike.models.ConvMessage.ParticipantInfoState;
 import com.bsb.hike.models.ConvMessage.State;
-import com.bsb.hike.models.Conversation.*;
+import com.bsb.hike.models.Conversation.BotConversation;
+import com.bsb.hike.models.Conversation.BroadcastConversation;
+import com.bsb.hike.models.Conversation.ConvInfo;
+import com.bsb.hike.models.Conversation.Conversation;
+import com.bsb.hike.models.Conversation.ConversationMetadata;
+import com.bsb.hike.models.Conversation.GroupConversation;
+import com.bsb.hike.models.Conversation.OneToNConvInfo;
+import com.bsb.hike.models.Conversation.OneToNConversation;
+import com.bsb.hike.models.Conversation.OneToNConversationMetadata;
+import com.bsb.hike.models.Conversation.OneToOneConversation;
+import com.bsb.hike.models.CustomStickerCategory;
+import com.bsb.hike.models.FileListItem;
+import com.bsb.hike.models.GroupParticipant;
+import com.bsb.hike.models.HikeFile;
 import com.bsb.hike.models.HikeFile.HikeFileType;
+import com.bsb.hike.models.HikeSharedFile;
+import com.bsb.hike.models.MessageEvent;
+import com.bsb.hike.models.MessageMetadata;
+import com.bsb.hike.models.Protip;
+import com.bsb.hike.models.StickerCategory;
 import com.bsb.hike.modules.contactmgr.ContactManager;
 import com.bsb.hike.modules.contactmgr.ConversationMsisdns;
 import com.bsb.hike.modules.contactmgr.GroupDetails;
 import com.bsb.hike.offline.OfflineUtils;
-import com.bsb.hike.platform.*;
+import com.bsb.hike.platform.ContentLove;
+import com.bsb.hike.platform.HikePlatformConstants;
+import com.bsb.hike.platform.PlatformMessageMetadata;
+import com.bsb.hike.platform.PlatformUtils;
+import com.bsb.hike.platform.WebMetadata;
 import com.bsb.hike.service.UpgradeIntentService;
 import com.bsb.hike.timeline.model.ActionsDataModel;
 import com.bsb.hike.timeline.model.ActionsDataModel.ActionTypes;
@@ -57,7 +90,14 @@ import com.bsb.hike.timeline.model.StatusMessage;
 import com.bsb.hike.timeline.model.StatusMessage.StatusMessageType;
 import com.bsb.hike.timeline.model.TimelineActions;
 import com.bsb.hike.timeline.view.TimelineActivity;
-import com.bsb.hike.utils.*;
+import com.bsb.hike.utils.ChatTheme;
+import com.bsb.hike.utils.HikeSharedPreferenceUtil;
+import com.bsb.hike.utils.Logger;
+import com.bsb.hike.utils.OneToNConversationUtils;
+import com.bsb.hike.utils.PairModified;
+import com.bsb.hike.utils.StealthModeManager;
+import com.bsb.hike.utils.StickerManager;
+import com.bsb.hike.utils.Utils;
 
 public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBConstants, HIKE_CONV_DB
 {
@@ -289,6 +329,10 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 
 		// to be aware of the users for whom db upgrade should not be done in future to fix AND-704
 		saveCurrentConvDbVersionToPrefs();
+		
+		String sqlIndex = "CREATE UNIQUE INDEX IF NOT EXISTS " + DBConstants.FEED_INDEX + " ON " + DBConstants.FEED_TABLE + " ( " + DBConstants.FEED_ACTION_ID + ", "
+				+ DBConstants.FEED_OBJECT_ID + ", " + DBConstants.FEED_ACTOR + " ) ";
+		db.execSQL(sqlIndex);
 	}
 
 	private void createIndexOverServerIdField(SQLiteDatabase db)
@@ -856,7 +900,7 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 				db.execSQL(alterTable);
 			}
 		}
-		
+
 		if (oldVersion < 47)
 		{
 			Long time = System.currentTimeMillis();
@@ -864,6 +908,25 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 			db.execSQL(getSortingIndexQuery()); //This index enables O(1) access for max sort id query, which will be used frequently
 			Logger.d("HikeConversationsDatabase", "Time taken to create indices for sortingId : " + (System.currentTimeMillis() - time));
 		}
+
+
+		if(oldVersion < 48)
+		{
+			try
+			{
+				String sqlIndex = "CREATE UNIQUE INDEX IF NOT EXISTS " + DBConstants.FEED_INDEX + " ON " + DBConstants.FEED_TABLE + " ( " + DBConstants.FEED_ACTION_ID + ", "
+						+ DBConstants.FEED_OBJECT_ID + ", " + DBConstants.FEED_ACTOR + " ) ";
+				db.execSQL(sqlIndex);
+			}
+			catch (SQLiteException sqe)
+			{
+				db.delete(DBConstants.FEED_TABLE, null, null);
+				String sqlIndex = "CREATE UNIQUE INDEX IF NOT EXISTS " + DBConstants.FEED_INDEX + " ON " + DBConstants.FEED_TABLE + " ( " + DBConstants.FEED_ACTION_ID + ", "
+						+ DBConstants.FEED_OBJECT_ID + ", " + DBConstants.FEED_ACTOR + " ) ";
+				db.execSQL(sqlIndex);
+			}
+		}
+
 
 	}
 
@@ -1021,7 +1084,7 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 	{
 		ContentValues values = new ContentValues();
 		values.put(DBConstants.TYPE, 0);
-		mDb.updateWithOnConflict(DBConstants.GROUP_MEMBERS_TABLE, values, MSISDN + "=?", new String[] { msisdn }, SQLiteDatabase.CONFLICT_REPLACE);		
+		mDb.updateWithOnConflict(DBConstants.GROUP_MEMBERS_TABLE, values, MSISDN + "=?", new String[]{msisdn}, SQLiteDatabase.CONFLICT_REPLACE);
 	}
 
 	/**
@@ -1729,28 +1792,48 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 	public int getUnreadActivityFeedCount()
 	{
 		String where = DBConstants.READ + " = 0 ";
-		int rowID = -1;
-		
+		int count = 0;
+
 		Cursor cursor = mDb.query(DBConstants.FEED_TABLE, null, where, null, null, null, null);
 
-		if(cursor != null)
+		if (cursor != null && cursor.moveToFirst())
 		{
-			rowID = cursor.getCount();
+			int columnIndex = cursor.getColumnIndex(DBConstants.FEED_ACTOR);
+			do
+			{
+				String msisdn = cursor.getString(columnIndex);
+				if (StealthModeManager.getInstance().isStealthMsisdn(msisdn) && !StealthModeManager.getInstance().isActive())
+				{
+					continue;
+				}
+				else
+				{
+					++count;
+				}
+			}
+			while (cursor.moveToNext());
 		}
-		return rowID;
+
+		return count;
 	}
 	
-	public boolean isAnyFeedEntryPresent()
-	{
-		String count = "SELECT count(*) FROM " + DBConstants.FEED_TABLE;
-		Cursor mcursor = mDb.rawQuery(count, null);
-		if(mcursor != null && mcursor.moveToFirst())
+	public boolean isAnyFeedEntryPresent() {
+		String count = "SELECT * FROM " + DBConstants.FEED_TABLE;
+		Cursor cursor = mDb.rawQuery(count, null);
+		if (cursor != null && cursor.moveToFirst())
 		{
-			int icount = mcursor.getInt(0);
-			if(icount > 0)
-			{
-				return true;
-			}
+			int columnIndex = cursor.getColumnIndex(DBConstants.FEED_ACTOR);
+			do {
+				String msisdn = cursor.getString(columnIndex);
+				if(StealthModeManager.getInstance().isStealthMsisdn(msisdn) && !StealthModeManager.getInstance().isActive())
+				{
+					continue;
+				}
+				else
+				{
+					return true;
+				}
+			} while (cursor.moveToNext());
 		}
 		return false;
 	}
@@ -1766,7 +1849,7 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 
 		String whereClause = DBConstants.FEED_OBJECT_ID + "=?";
 
-		int rowDeleted = mDb.delete(DBConstants.FEED_TABLE, whereClause, new String[] { mappedId });
+		int rowDeleted = mDb.delete(DBConstants.FEED_TABLE, whereClause, new String[]{mappedId});
 
 		if (rowDeleted != 0)
 		{
@@ -1833,19 +1916,63 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 		return isComplete;
 	}
 	
-	/**
-	 * 
-	 * @param currentPage 
-	 * @return
-	 */
-	public Cursor getActivityFeedsCursor()
+	public Cursor getActivityFeedsCursor(String[] msisdnList)
 	{
-		 String selectQuery = "SELECT  * FROM " + FEED_TABLE + " ft, "
+		StringBuilder msisdnSelection = null;
+		if (msisdnList != null)
+		{
+			msisdnSelection = new StringBuilder("(");
+			for (String msisdn : msisdnList)
+			{
+				msisdnSelection.append(DatabaseUtils.sqlEscapeString(msisdn) + ",");
+			}
+			msisdnSelection.replace(msisdnSelection.lastIndexOf(","), msisdnSelection.length(), ")");
+		}
+
+		StringBuilder selectQuery = new StringBuilder("SELECT  * FROM " + FEED_TABLE + " ft, "
 		            + STATUS_TABLE + " st " + " WHERE ft." + FEED_OBJECT_ID
-		            + " = " + "st." + STATUS_MAPPED_ID + " AND ft." 
-		            + FEED_OBJECT_TYPE + " = '" +  ActivityObjectTypes.STATUS_UPDATE.getTypeString()
-		            + "' ORDER BY ft." + FEED_TS + " DESC";
-		 return mDb.rawQuery(selectQuery, null);
+		            + " = " + "st." + STATUS_MAPPED_ID + " AND ft."
+		            + FEED_OBJECT_TYPE + " = '" +  ActivityObjectTypes.STATUS_UPDATE.getTypeString()+"'");
+
+		if (!TextUtils.isEmpty(msisdnSelection))
+		{
+			selectQuery.append(" AND "+"ft."+FEED_ACTOR + " IN " + msisdnSelection.toString());
+		}
+
+		selectQuery.append(" ORDER BY ft." + FEED_TS + " DESC");
+
+		return mDb.rawQuery(selectQuery.toString(), null);
+	}
+
+	public static String[] getTimelineFriendsMsisdn(String userMsisdn)
+	{
+		List<ContactInfo> friendsList = ContactManager.getInstance().getContactsOfFavoriteType(ContactInfo.FavoriteType.FRIEND, HikeConstants.BOTH_VALUE, userMsisdn);
+
+		Logger.d(HikeConstants.TIMELINE_LOGS, "list of friends from CM before filter" + friendsList);
+
+		ArrayList<String> msisdnList = new ArrayList<String>();
+
+		for (ContactInfo contactInfo : friendsList)
+		{
+			if (TextUtils.isEmpty(contactInfo.getMsisdn()))
+			{
+				continue;
+			}
+
+			if(StealthModeManager.getInstance().isStealthMsisdn(contactInfo.getMsisdn()) && !StealthModeManager.getInstance().isActive())
+			{
+				continue;
+			}
+
+			msisdnList.add(contactInfo.getMsisdn());
+		}
+
+		msisdnList.add(userMsisdn);
+
+		String[] friendMsisdns = new String[msisdnList.size()];
+		msisdnList.toArray(friendMsisdns);
+		Logger.d(HikeConstants.TIMELINE_LOGS, "list of friends after filter whose SU we are fetching " + friendMsisdns);
+		return friendMsisdns;
 	}
 	
 	/**
@@ -1982,15 +2109,15 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 		{
 			return;
 		}
-		mDb.update(DBConstants.SHARED_MEDIA_TABLE, contentValues, DBConstants.SERVER_ID + "=?", new String[] { String.valueOf(serverId) });
+		mDb.update(DBConstants.SHARED_MEDIA_TABLE, contentValues, DBConstants.SERVER_ID + "=?", new String[]{String.valueOf(serverId)});
 	}
 
 	public void updateConversationMetadata(String msisdn, com.bsb.hike.models.Conversation.ConversationMetadata metadata)
 	{
 		ContentValues contentValues = new ContentValues(1);
 		contentValues.put(DBConstants.CONVERSATION_METADATA, metadata.toString());
-		mDb.update(DBConstants.CONVERSATIONS_TABLE, contentValues, DBConstants.MSISDN + "=?", new String[] { msisdn });
-		HikeMessengerApp.getPubSub().publish(HikePubSub.CONV_META_DATA_UPDATED, new Pair<String, ConversationMetadata>(msisdn,metadata));
+		mDb.update(DBConstants.CONVERSATIONS_TABLE, contentValues, DBConstants.MSISDN + "=?", new String[]{msisdn});
+		HikeMessengerApp.getPubSub().publish(HikePubSub.CONV_META_DATA_UPDATED, new Pair<String, ConversationMetadata>(msisdn, metadata));
 	}
 
 	private void bindConversationInsert(SQLiteStatement insertStatement, ConvMessage conv,boolean bindForConvId)
@@ -2089,7 +2216,7 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 		}
 		ArrayList<ContactInfo> contacts= new ArrayList<ContactInfo>(1);
 		String msisdn = convMessages.get(0).getMsisdn();
-		contacts.add(new ContactInfo(msisdn, msisdn, null, null,!convMessages.get(0).isSMS()));
+		contacts.add(new ContactInfo(msisdn, msisdn, null, null, !convMessages.get(0).isSMS()));
 
 		return addConversations(convMessages, contacts,createConvIfNotExist);
 	}
@@ -2113,7 +2240,7 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 				// we don't need to re-insert it here
 				contentValues.put(DBConstants.MESSAGE_HASH, conv.createMessageHash());
 			}
-			mDb.update(DBConstants.MESSAGES_TABLE, contentValues, DBConstants.MESSAGE_ID + "=?", new String[] { Long.toString(conv.getMsgID()) });
+			mDb.update(DBConstants.MESSAGES_TABLE, contentValues, DBConstants.MESSAGE_ID + "=?", new String[]{Long.toString(conv.getMsgID())});
 
 		}
 		catch (Exception e)
@@ -2794,7 +2921,7 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 			mDb.beginTransaction();
 			mDb.delete(DBConstants.CONVERSATIONS_TABLE, DBConstants.MSISDN + "=?", new String[] { msisdn });
 			mDb.delete(DBConstants.MESSAGES_TABLE, DBConstants.MSISDN + "=?", new String[] { msisdn });
-			mDb.delete(DBConstants.SHARED_MEDIA_TABLE, DBConstants.MSISDN + "=?", new String[] { msisdn });
+			mDb.delete(DBConstants.SHARED_MEDIA_TABLE, DBConstants.MSISDN + "=?", new String[]{msisdn});
 			
 			if (OneToNConversationUtils.isOneToNConversation(msisdn))
 			{
@@ -3805,8 +3932,8 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 
 		try
 		{
-			c = mDb.query(DBConstants.CONVERSATIONS_TABLE, new String[] { DBConstants.MESSAGE, DBConstants.MSG_STATUS, DBConstants.LAST_MESSAGE_TIMESTAMP, DBConstants.MAPPED_MSG_ID,
-					DBConstants.MESSAGE_ID, DBConstants.MESSAGE_METADATA, DBConstants.GROUP_PARTICIPANT }, DBConstants.MSISDN + "=?", new String[] { msisdn }, null, null, null);
+			c = mDb.query(DBConstants.CONVERSATIONS_TABLE, new String[]{DBConstants.MESSAGE, DBConstants.MSG_STATUS, DBConstants.LAST_MESSAGE_TIMESTAMP, DBConstants.MAPPED_MSG_ID,
+					DBConstants.MESSAGE_ID, DBConstants.MESSAGE_METADATA, DBConstants.GROUP_PARTICIPANT}, DBConstants.MSISDN + "=?", new String[]{msisdn}, null, null, null);
 
 			final int msgColumn = c.getColumnIndex(DBConstants.MESSAGE);
 			final int msgStatusColumn = c.getColumnIndex(DBConstants.MSG_STATUS);
@@ -4172,7 +4299,7 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 		contentValues.put(DBConstants.UNREAD_COUNT, 0);
 		contentValues.put(DBConstants.MSG_STATUS, State.RECEIVED_READ.ordinal());
 
-		mDb.update(DBConstants.CONVERSATIONS_TABLE, contentValues, DBConstants.MSISDN + "=?", new String[] { msisdn });
+		mDb.update(DBConstants.CONVERSATIONS_TABLE, contentValues, DBConstants.MSISDN + "=?", new String[]{msisdn});
 	}
 
 	private void deleteMessageFromConversation(String msisdn)
@@ -4915,8 +5042,13 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 	{
 		return getStatusMessages(timelineUpdatesOnly, -1, -1, msisdnList);
 	}
-	
+
 	public List<StatusMessage> getStatusMessages(boolean timelineUpdatesOnly, int limit, int[] types)
+	{
+		return getStatusMessages(timelineUpdatesOnly,limit,types,false);
+	}
+
+	public List<StatusMessage> getStatusMessages(boolean timelineUpdatesOnly, int limit, int[] types, boolean getOnlyStealthStatus)
 	{
 		if (types == null || types.length == 0)
 		{
@@ -4972,6 +5104,14 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 			while (c.moveToNext())
 			{
 				String msisdn = c.getString(msisdnIdx);
+
+				if(getOnlyStealthStatus)
+				{
+					if(!StealthModeManager.getInstance().isStealthMsisdn(msisdn))
+					{
+						continue;
+					}
+				}
 
 				StatusMessage statusMessage = new StatusMessage(c.getLong(idIdx), c.getString(mappedIdIdx), msisdn, null, c.getString(textIdx),
 						StatusMessageType.values()[c.getInt(typeIdx)], c.getLong(tsIdx), c.getInt(moodIdIdx), c.getInt(timeOfDayIdx), c.getString(fileKeyIdx));
@@ -5461,10 +5601,12 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 		updateStickerCategoryData(categoryId, null, totalNum, -1);
 	}
 
-	public void removeStickerCategory(String categoryId)
+	public void removeStickerCategory(String categoryId, boolean removeFromShopTable)
 	{
 		mDb.delete(DBConstants.STICKER_CATEGORIES_TABLE, DBConstants._ID + "=?", new String[] { categoryId });
-		mDb.delete(DBConstants.STICKER_SHOP_TABLE, DBConstants._ID + "=?", new String[] { categoryId });
+		if (removeFromShopTable) {
+			mDb.delete(DBConstants.STICKER_SHOP_TABLE, DBConstants._ID + "=?", new String[]{categoryId});
+		}
 	}
 
 	public void updateStickerCategoryData(String categoryId, Boolean updateAvailable, int totalStickerCount, int categorySize)
@@ -5674,7 +5816,7 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 
 	public void deleteProtip(String mappedId)
 	{
-		mDb.delete(DBConstants.PROTIP_TABLE, DBConstants.PROTIP_MAPPED_ID + "=?", new String[] { mappedId });
+		mDb.delete(DBConstants.PROTIP_TABLE, DBConstants.PROTIP_MAPPED_ID + "=?", new String[]{mappedId});
 	}
 
 	public void incrementUnreadCounter(String msisdn)
@@ -5809,7 +5951,22 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 					metadata.getJSON().optJSONArray(HikeConstants.FILES).optJSONObject(0).remove(HikeConstants.THUMBNAIL);
 				}
 			}
-			contentValues.put(DBConstants.MESSAGE_METADATA, metadata.getJSON().optJSONArray(HikeConstants.FILES).optJSONObject(0).toString());
+			JSONObject jsonObj = metadata.getJSON().optJSONArray(HikeConstants.FILES).optJSONObject(0);
+
+			String caption = metadata.getJSON().optString(HikeConstants.CAPTION);
+			if(!TextUtils.isEmpty(caption))
+			{
+				try
+				{
+					jsonObj.put(HikeConstants.CAPTION, caption);
+				}
+				catch (JSONException e)
+				{
+					e.printStackTrace();
+				}
+			}
+
+			contentValues.put(DBConstants.MESSAGE_METADATA, jsonObj.toString());
 
 			if (thumbnailString != null && !TextUtils.isEmpty(thumbnailString))
 			{
@@ -6952,6 +7109,19 @@ public class HikeConversationsDatabase extends SQLiteOpenHelper implements DBCon
 			{
 				c.close();
 			}
+		}
+	}
+
+	public void deleteActivityFeedForMsisdn(List<String> stealthMsisdns)
+	{
+		if(Utils.isEmpty(stealthMsisdns))
+		{
+			return;
+		}
+
+		for(String msisdn: stealthMsisdns)
+		{
+			mDb.delete(DBConstants.FEED_TABLE, DBConstants.FEED_ACTOR + "=?", new String[] {msisdn});
 		}
 	}
